// MantenedoresPerfilCliente.Angular/src/app/shared/models/filtro-dto.ts
export class FiltroDto {
  id: number;
  codigo: number;
  descripcion: string;
  universoId: number;
  universoDescripcion: string;
  estadoFiltroId: number;
  estadoFiltroDescripcion: string;
  perfilId: number;
  perfilDescripcion: string;
  orden: number;
  corte1: number;
  corte2: number;
  monto1: number;
  monto2: number;
}